proxy: move isLocalhost() impl HTTPProxy and do not rely on net.lookupStaticHost
Replace net.lookupStaticHost() with localhost aliases loaded on start.
